Fixed a race condition causing EOF to be returned too early:
If input was received just before global EOF, it was possible for the
input stream receiving the input to not re-acquire the lock (line 94 in
MultiplexedInputStream) and handle the received data before the global
EOF caused the stream-specific state to be set to EOF as well. In this
case, the received data would be lost and EOF (-1) would be returned
from read(...).